shopping-sync/requirements.txt
caldav>=1.3.0
icalendar>=5.0.0
playwright>=1.40.0
python-dotenv>=1.0.0
